Add test case for edit detail functionality
1.Add test case for edit details functionality,
2.create dist/ files for deployment
3.edit readme with testing changes in the project
