+++ .eslintignore
@@ -1,1 +1,3 @@
 coverage/
+__tests__/
+__mocks__/
